Added constructor for containers
enumerable can now be created directly from a
"container" - anything on which std::begin and
std::end would work. This disambiguates with
the constructor that accepts next delegate.

It is also no longer possible to copy the
container when calling enumerate_container.
This allows us to iterate containers of objects
that cannot be copied (or moved). If you want
the enumerable to store the container, move it.