Added size() method to enumerable
This is powered by an optional size delegate. If not provided we
use std::distance on the enumerable's iterators, which could be
slow, but is better than nothing. A method called has_fast_size()
can be used to validate whether we can quickly get the size.
Reworked some of the LINQ operators to use this.