Take out enumerable::const_iterator
Having iterator and const_iterator in enumerable did
not make much sense since enumerable is in itself a
front for a sequence. It would've been trivial to copy
the enumerable to access the non-const methods.
So now enumerable only has an iterator and begin
and end are both const methods. If you want const,
use enumerable<const foo> (added a new as_const
method to simplify this).